// src/app/models/employee/employee.model.ts
export class Employee {
    id: number;
    name: string;
    gender: string;
    email?: string;
    phoneNumber?: number;
    // password: string;
    // confirmPassowrd: string;
    contactPreference: string;
    dateOfBirth: Date;
    departments: string;
    isActive: boolean;
    photoPath?: string;
   }
  